+++ src/api/axiosInstance.ts
@@ -2,7 +2,7 @@ import axios, { type AxiosRequestConfig } from 'axios'
 
 import { getOidc } from '@/oidc'
 
-const axiosInstance = axios.create({
+export const axiosInstance = axios.create({
   baseURL: import.meta.env.VITE_API_URL,
 })