+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Authentication:
   - retrive scopes for jwt token
   - new env variable: `VITE_OIDC_SCOPES` (default value `"profile,roles"`)
+  - added authentification check on visualization page
 
 ## [2.7.0](https://github.com/InseeFr/stromae-dsfr/releases/tag/2.7.0) - 2026-04-14